Rename :docs-link etc to :docs_link
This appears to fix the error in [issue
3577](https://github.com/emacs-lsp/lsp-mode/issues/3577) for me: as I
made these edits in my local straight copy of lsp-mode, I saw the error
message (which was appearing when I called 'revert-buffer in a
non-terraform buffer for which lsp-mode was configured) move from
reporting

```
Error running timer: (error "Eager macro-expansion failure: (error
\"Unknown key: :docs-link.  Available keys: (:name :docs_link :version
:source_type :dependent_modules)\")")
```

to

```
Error running timer: (error "Eager macro-expansion failure: (error
\"Unknown key: :source-type.  Available keys: (:name :docs_link :version
:source_type :dependent_modules)\")")
```

and so on.

Interestingly, flycheck is reporting the exact opposite message now,
claiming that “:docs_link” is an unknown key and “:docs-link” should be
available.

Not sure what’s going on!